pyproject.toml
[project]
name = "hotel-spider"
version = "0.1.0"
description = "Hotel competitor price monitoring MVP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "pydantic-settings>=2.10.0,<3.0.0",
    "sqlalchemy>=2.0.40,<3.0.0",
    "uvicorn>=0.35.0,<1.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.4.0,<9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hotel_spider"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
